% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_blocks.R
\name{actions_block}
\alias{actions_block}
\title{Action Block}
\usage{
actions_block(elements, block_id = NULL, type = "actions")
}
\arguments{
\item{elements}{A list of interactive element objects - buttons, select menus, overflow menus, or date pickers. There is a maximum of 5 elements in each action block.}

\item{block_id}{A string acting as a unique identifier for a block. If not specified, a block_id will be generated. You can use this block_id when you receive an interaction payload to identify the source of the action. Maximum length for this field is 255 characters. block_id should be unique for each message and each iteration of a message. If a message is updated, use a new block_id.}

\item{type}{The type of block. For an actions block, type is always actions.}
}
\value{
A \href{https://api.slack.com/reference/block-kit/blocks#actions}{Slack Action Block}
}
\description{
A block that is used to hold \href{https://api.slack.com/reference/block-kit/block-elements}{interactive elements}.
}
