% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messaging.R
\name{attachment_field_object}
\alias{attachment_field_object}
\title{Attachment Field Object}
\usage{
attachment_field_object(title = NULL, value = NULL, short = NULL)
}
\arguments{
\item{title}{Shown as a bold heading displayed in the field object. It cannot contain markup and will be escaped for you.}

\item{value}{The text value displayed in the field object. It can be formatted as plain text, or with mrkdwn by using the mrkdwn_in option above.}

\item{short}{Indicates whether the field object is short enough to be displayed side-by-side with other field objects. Defaults to false.}
}
\value{
An attachment field object
}
\description{
Create an attachment field object
}
\seealso{
\url{https://api.slack.com/reference/messaging/attachments#fields}

Other Messaging: 
\code{\link{attachment_object}()},
\code{\link{message_object}()}
}
\concept{Messaging}
