% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messaging.R
\name{attachment_object}
\alias{attachment_object}
\title{Message Attachment Object}
\usage{
attachment_object(
  author_icon = NULL,
  author_link = NULL,
  author_name = NULL,
  blocks = NULL,
  color = NULL,
  fallback = NULL,
  fields = NULL,
  footer = NULL,
  footer_icon = NULL,
  image_url = NULL,
  mrkdwn_in = NULL,
  pretext = NULL,
  text = NULL,
  thumb_url = NULL,
  title = NULL,
  title_link = NULL,
  ts = NULL
)
}
\arguments{
\item{author_icon}{A valid URL that displays a small 16px by 16px image to the left of the author_name text. Will only work if author_name is present.}

\item{author_link}{A valid URL that will hyperlink the author_name text. Will only work if author_name is present.}

\item{author_name}{Small text used to display the author's name.}

\item{blocks}{A list of @family{Blocks} in the same format as described in the \href{https://api.slack.com/block-kit/building}{building blocks guide}.}

\item{color}{Changes the color of the border on the left side of this attachment from the default gray. Can either be one of good (green), warning (yellow), danger (red), or any hex color code (eg. #439FE0)}

\item{fallback}{A plain text summary of the attachment used in clients that don't show formatted text (eg. IRC, mobile notifications).}

\item{fields}{A list of field objects that get displayed in a table-like way (See the example above). For best results, include no more than 2-3 field objects.}

\item{footer}{Some brief text to help contextualize and identify an attachment. Limited to 300 characters, and may be truncated further when displayed to users in environments with limited screen real estate.}

\item{footer_icon}{A valid URL to an image file that will be displayed beside the footer text. Will only work if author_name is present. We'll render what you provide at 16px by 16px. It's best to use an image that is similarly sized.}

\item{image_url}{A valid URL to an image file that will be displayed at the bottom of the attachment. We support GIF, JPEG, PNG, and BMP formats.Large images will be resized to a maximum width of 360px or a maximum height of 500px, while still maintaining the original aspect ratio. Cannot be used with thumb_url.}

\item{mrkdwn_in}{A list of field names that should be formatted by mrkdwn syntax.}

\item{pretext}{Text that appears above the message attachment block. It can be formatted as plain text, or with mrkdwn by including it in the mrkdwn_in field.}

\item{text}{The main body text of the attachment. It can be formatted as plain text, or with mrkdwn by including it in the mrkdwn_in field. The content will automatically collapse if it contains 700+ characters or 5+ linebreaks, and will display a "Show more..." link to expand the content.}

\item{thumb_url}{A valid URL to an image file that will be displayed as a thumbnail on the right side of a message attachment. We currently support the following formats: GIF, JPEG, PNG, and BMP.The thumbnail's longest dimension will be scaled down to 75px while maintaining the aspect ratio of the image. The filesize of the image must also be less than 500 KB.For best results, please use images that are already 75px by 75px.}

\item{title}{Large title text near the top of the attachment.}

\item{title_link}{A valid URL that turns the title text into a hyperlink.}

\item{ts}{An integer Unix timestamp that is used to related your attachment to a specific time. The attachment will display the additional timestamp value as part of the attachment's footer.Your message's timestamp will be displayed in varying ways, depending on how far in the past or future it is, relative to the present. Form factors, like mobile versus desktop may also transform its rendered appearance.}
}
\value{
A message attachment object
}
\description{
This feature is a legacy part of messaging functionality for Slack apps. We recommend you stick with layout blocks, but if you still want to use attachments, read our caveats.
}
\seealso{
\url{https://api.slack.com/reference/messaging/payload}

Other Messaging: 
\code{\link{attachment_field_object}()},
\code{\link{message_object}()}
}
\concept{Messaging}
