% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_elements.R
\name{button.element}
\alias{button.element}
\title{Create a Button Element}
\usage{
button.element(
  type = "button",
  text,
  action_id = NULL,
  url = NULL,
  value = NULL,
  style = NULL,
  confirm = NULL
)
}
\arguments{
\item{type}{The type of element. In this case type is always button.}

\item{text}{A \code{\link{text.object}} that defines the button's text. Can only be of type: plain_text. Maximum length for the text in this field is 75 characters.}

\item{action_id}{An identifier for this action. You can use this when you receive an interaction payload to identify the source of the action. Should be unique among all other action_ids in the containing block. Maximum length for this field is 255 characters.}

\item{url}{A URL to load in the user's browser when the button is clicked. Maximum length for this field is 3000 characters. If you're using url, you'll still receive an \href{https://api.slack.com/interactivity/handling#payloads}{interaction payload} and will need to send an \href{https://api.slack.com/interactivity/handling#acknowledgment_response}{acknowledgment response}.}

\item{value}{The value to send along with the interaction payload. Maximum length for this field is 2000 characters.}

\item{style}{Decorates buttons with alternative visual color schemes. Use this option with restraint.primary gives buttons a green outline and text, ideal for affirmation or confirmation actions. primary should only be used for one button within a set. danger gives buttons a red outline and text, and should be used when the action is destructive. Use danger even more sparingly than primary. If you don't include this field, the default button style will be used.}

\item{confirm}{A \code{\link{confirm.object}} that defines an optional confirmation dialog after the button is clicked.}
}
\value{
A button element
}
\description{
An interactive component that inserts a button. The button can be a trigger for anything from opening a simple link to starting a complex workflow.
}
\seealso{
\url{https://api.slack.com/reference/block-kit/block-elements#button}
}
