% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_elements.R
\name{channels_multi_select_menu}
\alias{channels_multi_select_menu}
\title{Public Channels List Multi-Select Menu Element}
\usage{
channels_multi_select_menu(
  placeholder,
  action_id,
  initial_channels = NULL,
  confirm = NULL,
  max_selected_items = NULL
)
}
\arguments{
\item{placeholder}{A plain_text only \code{\link{text_object}} that defines the placeholder text shown on the menu. Maximum length for the text in this field is 150 characters.}

\item{action_id}{An identifier for the action triggered when a menu option is selected. You can use this when you receive an interaction payload to \href{https://api.slack.com/messaging/interactivity/enabling#understanding_payloads}{identify the source of the action}. Should be unique among all other action_ids in the containing block. Maximum length for this field is 255 characters.}

\item{initial_channels}{A list of one or more IDs of any valid public channel to be pre-selected when the menu loads.}

\item{confirm}{A \code{\link{confirm_object}} that defines an optional confirmation dialog that appears before the multi-select choices are submitted.}

\item{max_selected_items}{Specifies the maximum number of items that can be selected in the menu. Minimum number is 1.}
}
\description{
This multi-select menu will populate its options with a list of public channels visible to the current user in the active workspace. A multi-select menu allows a user to select multiple items from a list of options. Just like regular select menus, multi-select menus also include type-ahead functionality, where a user can type a part or all of an option string to filter the list. To use interactive components, you will need to make some changes to prepare your app. Read our \href{https://api.slack.com/interactivity/handling}{guide to enabling interactivity}. Works with block types: Section Input
}
\seealso{
\url{https://api.slack.com/reference/block-kit/block-elements#channel_multi_select_menu}

Other Elements: 
\code{\link{button_element}()},
\code{\link{checkbox_element}()},
\code{\link{conversations_multi_select_menu}()},
\code{\link{datepicker_element}()},
\code{\link{external_multi_select_menu}()},
\code{\link{image_element}()},
\code{\link{overflow_menu_element}()},
\code{\link{plain_text_input_element}()},
\code{\link{radio_button_group_element}()},
\code{\link{static_multi_select_menu}()},
\code{\link{users_multi_select_menu}()}
}
\concept{Elements}
