% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_elements.R
\name{channels_select_menu}
\alias{channels_select_menu}
\title{Public Channels List Select Menu Element}
\usage{
channels_select_menu(
  placeholder,
  action_id,
  initial_channel = NULL,
  confirm = NULL
)
}
\arguments{
\item{placeholder}{A plain_text only \code{\link{text_object}} that defines the placeholder text shown on the menu. Maximum length for the text in this field is 150 characters.}

\item{action_id}{An identifier for the action triggered when a menu option is selected. You can use this when you receive an interaction payload to \href{https://api.slack.com/messaging/interactivity/enabling#understanding_payloads}{identify the source of the action}. Should be unique among all other action_ids in the containing block. Maximum length for this field is 255 characters.}

\item{initial_channel}{The ID of any valid public channel to be pre-selected when the menu loads.}

\item{confirm}{A \code{\link{confirm_object}} that defines an optional confirmation dialog that appears before the select choices are submitted.}

\item{response_url_enabled}{This field only works with menus in \code{\link{input_block}}s in \href{https://api.slack.com/surfaces/modals}{modals}. When set to true, the \href{https://api.slack.com/reference/interaction-payloads/views#view_submission}{view_submission payload} from the menu's parent view will contain a response_url. This response_url can be used for \href{https://api.slack.com/interactivity/handling#message_responses}{message responses}. The target channel for the message will be determined by the value of this select menu.}
}
\description{
This multi-select menu will populate its options with a list of public channels visible to the current user in the active workspace. A select menu, just as with a standard HTML \verb{<select>} tag, creates a drop down menu with a list of options for a user to choose. The select menu also includes type-ahead functionality, where a user can type a part or all of an option string to filter the list. To use interactive components, you will need to make some changes to prepare your app. Read our \href{https://api.slack.com/interactivity/handling}{guide to enabling interactivity}. Works with block types: Section Actions Input
}
\seealso{
\url{https://api.slack.com/reference/block-kit/block-elements#channel_select}

Other Elements: 
\code{\link{button_element}()},
\code{\link{channels_multi_select_menu}()},
\code{\link{checkbox_element}()},
\code{\link{conversations_multi_select_menu}()},
\code{\link{conversations_select_menu}()},
\code{\link{datepicker_element}()},
\code{\link{external_multi_select_menu}()},
\code{\link{external_select_menu}()},
\code{\link{image_element}()},
\code{\link{overflow_menu_element}()},
\code{\link{plain_text_input_element}()},
\code{\link{radio_button_group_element}()},
\code{\link{static_multi_select_menu}()},
\code{\link{static_select_menu}()},
\code{\link{timepicker_element}()},
\code{\link{users_multi_select_menu}()},
\code{\link{users_select_menu}()}
}
\concept{Elements}
