% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_objects.R
\name{confirm_object}
\alias{confirm_object}
\title{Confirmation dialog object}
\usage{
confirm_object(title, text, confirm, deny, style = NULL)
}
\arguments{
\item{title}{A plain_text-only \code{\link{text_object}} that defines the dialog's title. Maximum length for this field is 100 characters.}

\item{text}{A \code{\link{text_object}} that defines the explanatory text that appears in the confirm dialog. Maximum length for the text in this field is 300 characters.}

\item{confirm}{A plain_text-only \code{\link{text_object}} to define the text of the button that confirms the action. Maximum length for the text in this field is 30 characters.}

\item{deny}{A plain_text-only \code{\link{text_object}} to define the text of the button that cancels the action. Maximum length for the text in this field is 30 characters.}

\item{style}{Defines the color scheme applied to the confirm button. A value of danger will display the button with a red background on desktop, or red text on mobile. A value of primary will display the button with a green background on desktop, or blue text on mobile. If this field is not provided, the default value will be primary.}
}
\value{
A Slack Confirmation Object
}
\description{
An object that defines a dialog that provides a confirmation step to any interactive element. This dialog will ask the user to confirm their action by offering a confirm and deny buttons.
}
\seealso{
\url{https://api.slack.com/reference/block-kit/composition-objects#confirm}
}
