% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_blocks.R
\name{context_block}
\alias{context_block}
\title{Context Block}
\usage{
context_block(elements, block_id = NULL)
}
\arguments{
\item{elements}{A list of \code{\link{image_element}}s and \code{\link{text_object}}s. Maximum number of items is 10.}

\item{block_id}{A string acting as a unique identifier for a block. If not specified, one will be generated. Maximum length for this field is 255 characters. block_id should be unique for each message and each iteration of a message. If a message is updated, use a new block_id.}
}
\value{
A \href{https://api.slack.com/reference/block-kit/blocks#context}{Slack Context Block}
}
\description{
Displays message context, which can include both images and text.
}
