% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_objects.R
\name{conversation_list_filter}
\alias{conversation_list_filter}
\title{Filter Object for Conversation Lists}
\usage{
conversation_list_filter(
  include = NULL,
  exclude_external_shared_channels = NULL,
  exclude_bot_users = NULL
)
}
\arguments{
\item{include}{Indicates which type of conversations should be included in the list. When this field is provided, any conversations that do not match will be excluded. You should provide an array of strings from the following options: im, mpim, private, and public. The array cannot be empty.}

\item{exclude_external_shared_channels}{Indicates whether to exclude external \href{https://api.slack.com/enterprise/shared-channels}{shared channels} from conversation lists. Defaults to false.}

\item{exclude_bot_users}{Indicates whether to exclude bot users from conversation lists. Defaults to false.}
}
\description{
Provides a way to filter the list of options in a \code{\link{conversations_select_menu}} or \code{\link{conversations_multi_select_menu}}. Please note that while none of the fields above are individually required, you must supply at least one of these fields.
}
\seealso{
\url{https://api.slack.com/reference/block-kit/composition-objects#filter_conversations}

Other Composition: 
\code{\link{confirm_object}()},
\code{\link{dispaction_action_configuration}()},
\code{\link{option_group}()},
\code{\link{option_object}()},
\code{\link{text_object}()}
}
\concept{Composition}
