% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversations.R
\name{conversations_open}
\alias{conversations_open}
\title{Opens or resumes a direct message or multi-person direct message.}
\usage{
conversations_open(
  token = Sys.getenv("SLACK_TOKEN"),
  channel = NULL,
  return_im = NULL,
  users = NULL,
  return_response = F
)
}
\arguments{
\item{token}{Authentication token bearing required scopes. Tokens should be passed as an HTTP Authorization header or alternatively, as a POST parameter.}

\item{channel}{Resume a conversation by supplying an im or mpim's ID. Or provide the users field instead.}

\item{return_im}{Boolean, indicates you want the full IM channel definition in the response.}

\item{users}{Comma separated lists of users. If only one user is included, this creates a 1:1 DM. The ordering of the users is preserved whenever a multi-person direct message is returned. Supply a channel when not supplying users.}
}
\description{
Opens or resumes a direct message or multi-person direct message.
}
\section{Details}{

This method supports application/json via HTTP POST. Present your token in your request's Authorization header. Learn more.
}

