% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_objects.R
\name{dispaction_action_configuration}
\alias{dispaction_action_configuration}
\title{Dispatch Action Configuration}
\usage{
dispaction_action_configuration(trigger_actions_on)
}
\arguments{
\item{trigger_actions_on}{A list of interaction types that you would like to receive a \href{https://api.slack.com/reference/interaction-payloads/block-actions}{block_actions payload} for. Should be one or both of:
\itemize{
\item{on_enter_pressed - payload is dispatched when user presses the enter key while the input is in focus. Hint text will appear underneath the input explaining to the user to press enter to submit.}
\item{on_character_entered - payload is dispatched when a character is entered (or removed) in the input.}
}}
}
\description{
Determines when a \code{\link{plain_text_input_element}} will return a block_actions \href{https://api.slack.com/reference/interaction-payloads/block-actions}{interaction payload}.
}
\seealso{
\url{https://api.slack.com/reference/block-kit/composition-objects#dispatch_action_config}

Other Composition: 
\code{\link{confirm_object}()},
\code{\link{conversation_list_filter}()},
\code{\link{option_group}()},
\code{\link{option_object}()},
\code{\link{text_object}()}
}
\concept{Composition}
