% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_blocks.R
\name{divider_block}
\alias{divider_block}
\title{Divider Block}
\usage{
divider_block(block_id = NULL)
}
\arguments{
\item{block_id}{A string acting as a unique identifier for a block. If not specified, one will be generated. Maximum length for this field is 255 characters. block_id should be unique for each message and each iteration of a message. If a message is updated, use a new block_id.}
}
\description{
A content divider, like an \verb{<hr>}, to split up different blocks inside of a message. The divider block is nice and neat, requiring only a type. Available in surfaces: Modals Messages Home tabs
}
\seealso{
\url{https://api.slack.com/reference/block-kit/blocks#divider}

Other Blocks: 
\code{\link{actions_block}()},
\code{\link{context_block}()},
\code{\link{file_block}()},
\code{\link{header_block}()},
\code{\link{image_block}()},
\code{\link{input_block}()},
\code{\link{section_block}()}
}
\concept{Blocks}
