% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_blocks.R
\name{file_block}
\alias{file_block}
\title{File Block}
\usage{
file_block(external_id, block_id = NULL)
}
\arguments{
\item{external_id}{The external unique ID for this file.}

\item{block_id}{A string acting as a unique identifier for a block. If not specified, one will be generated. Maximum length for this field is 255 characters. block_id should be unique for each message and each iteration of a message. If a message is updated, use a new block_id.}
}
\description{
Displays a remote file. You can't add this block to app surfaces directly, but it will show up when retrieving messages that contain remote files. If you want to add remote files to messages, \href{https://api.slack.com/messaging/files/remote}{follow our guide}. Appears in surfaces: Messages
}
\seealso{
\url{https://api.slack.com/reference/block-kit/blocks#file}

Other Blocks: 
\code{\link{actions_block}()},
\code{\link{context_block}()},
\code{\link{divider_block}()},
\code{\link{header_block}()},
\code{\link{image_block}()},
\code{\link{input_block}()},
\code{\link{section_block}()}
}
\concept{Blocks}
