% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{files_upload}
\alias{files_upload}
\title{Uploads or creates a file.}
\usage{
files_upload(
  token = Sys.getenv("SLACK_TOKEN"),
  channels = NULL,
  content = NULL,
  file = NULL,
  filename = NULL,
  filetype = NULL,
  initial_comment = NULL,
  thread_ts = NULL,
  title = NULL,
  return_response = F
)
}
\arguments{
\item{token}{Authentication token bearing required scopes. Tokens should be passed as an HTTP Authorization header or alternatively, as a POST parameter.}

\item{channels}{Comma-separated list of channel names or IDs where the file will be shared.}

\item{content}{File contents via a POST variable. If omitting this parameter, you must provide a file.}

\item{file}{File contents via multipart/form-data. If omitting this parameter, you must submit content.}

\item{filename}{Filename of file.}

\item{filetype}{A \href{https://api.slack.com/types/file#file_types}{file type} identifier.}

\item{initial_comment}{The message text introducing the file in specified channels.}

\item{title}{Title of file.}

\item{return_response}{Whether or not to return the API call response as opposed to the response body. Defaults to FALSE (return response body)}

\item{Provide}{another message's ts value to upload this file as a reply. Never use a reply's ts value; use its parent instead.}
}
\value{
The details of the uploaded file.
}
\description{
Uploads or creates a file.
}
\seealso{
\url{https://api.slack.com/methods/files.upload}
}
\concept{Files}
