% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_blocks.R
\name{header_block}
\alias{header_block}
\title{Header Block}
\usage{
header_block(text, block_id = NULL)
}
\arguments{
\item{text}{The text for the block, in the form of a plain_text \code{\link{text_object}}. Maximum length for the text in this field is 150 characters.}

\item{block_id}{A string acting as a unique identifier for a block. If not specified, one will be generated. Maximum length for this field is 255 characters. block_id should be unique for each message and each iteration of a message. If a message is updated, use a new block_id.}
}
\description{
A header is a plain-text block that displays in a larger, bold font. Use it to delineate between different groups of content in your app's surfaces. Available in surfaces: Modals Messages Home tabs
}
\seealso{
\url{https://api.slack.com/reference/block-kit/blocks#header}

Other Blocks: 
\code{\link{actions_block}()},
\code{\link{context_block}()},
\code{\link{divider_block}()},
\code{\link{file_block}()},
\code{\link{image_block}()},
\code{\link{input_block}()},
\code{\link{section_block}()}
}
\concept{Blocks}
