% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_blocks.R
\name{image_block}
\alias{image_block}
\title{Image Block}
\usage{
image_block(image_url, alt_text, title = NULL, block_id = NULL)
}
\arguments{
\item{image_url}{The URL of the image to be displayed. Maximum length for this field is 3000 characters.}

\item{alt_text}{A plain-text summary of the image. This should not contain any markup. Maximum length for this field is 2000 characters.}

\item{title}{An optional title for the image in the form of a \code{\link{text_object}} that can only be of type: plain_text. Maximum length for the text in this field is 2000 characters.}

\item{block_id}{A string acting as a unique identifier for a block. If not specified, one will be generated. Maximum length for this field is 255 characters. block_id should be unique for each message and each iteration of a message. If a message is updated, use a new block_id.}
}
\description{
A simple image block, designed to make those cat photos really pop. Available in surfaces: Modals Messages Home tabs
}
\seealso{
https://api.slack.com/reference/block-kit/blocks#image

Other Blocks: 
\code{\link{actions_block}()},
\code{\link{context_block}()},
\code{\link{divider_block}()},
\code{\link{file_block}()},
\code{\link{header_block}()},
\code{\link{input_block}()},
\code{\link{section_block}()}
}
\concept{Blocks}
