% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_blocks.R
\name{input_block}
\alias{input_block}
\title{Input Block}
\usage{
input_block(
  label,
  element,
  dispatch_action = NULL,
  block_id = NULL,
  hint = NULL,
  option = NULL
)
}
\arguments{
\item{label}{A label that appears above an input element in the form of a \code{\link{text_object}} that must have type of plain_text. Maximum length for the text in this field is 2000 characters.}

\item{element}{A \code{\link{plain_text_input_element}}, \code{\link{checkbox_element}}, \code{\link{radio_button_element}}, \code{\link{select_menu_element}}, \code{\link{multi_select_menu_element}}, or \code{\link{datepicker_element}}.}

\item{dispatch_action}{A boolean that indicates whether or not the use of elements in this block should dispatch a block_actions payload. Defaults to false.}

\item{block_id}{A string acting as a unique identifier for a block. If not specified, one will be generated. Maximum length for this field is 255 characters. block_id should be unique for each message and each iteration of a message. If a message is updated, use a new block_id.}

\item{hint}{An optional hint that appears below an input element in a lighter grey. It must be a \code{\link{text_object}} with a type of plain_text. Maximum length for the text in this field is 2000 characters.}

\item{optional}{A boolean that indicates whether the input element may be empty when a user submits the modal. Defaults to false.}
}
\description{
A block that collects information from users - it can hold a \code{\link{plain_text_input_element}}, a \code{\link{checkbox_element}}, a \code{\link{radio_button_element}}, a \code{\link{select_menu_element}}, a \code{\link{multi_select_menu_element}}, or a \code{\link{datepicker_element}}. Read our guides to collecting input \href{https://api.slack.com/surfaces/modals#gathering_input}{in modals} or \href{https://api.slack.com/surfaces/tabs/using#gathering_input}{in Home tabs} to learn how input blocks pass information to your app. Available in surfaces: Modals Home tabs
}
\seealso{
https://api.slack.com/reference/block-kit/blocks#input

Other Blocks: 
\code{\link{actions_block}()},
\code{\link{context_block}()},
\code{\link{divider_block}()},
\code{\link{file_block}()},
\code{\link{header_block}()},
\code{\link{image_block}()},
\code{\link{section_block}()}
}
\concept{Blocks}
