% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messaging.R
\name{message_object}
\alias{message_object}
\title{Message Object}
\usage{
message_object(
  text,
  blocks = NULL,
  attachments = NULL,
  thread_ts = NULL,
  mrkdown = NULL
)
}
\arguments{
\item{text}{The usage of this field changes depending on whether you're using blocks or not. If you are, this is used as a fallback string to display in notifications. If you aren't, this is the main body text of the message. It can be formatted as plain text, or with mrkdwn. This field is not enforced as required when using blocks, however it is highly recommended that you include it as the aforementioned fallback.}

\item{blocks}{An array of @family{Blocks} in the same format as described in the \href{https://api.slack.com/block-kit/building}{building blocks guide}.}

\item{attachments}{An array of legacy secondary attachments. We recommend you use blocks instead.}

\item{The}{ID of another un-threaded message to reply to.}

\item{mrkdwn}{Determines whether the text field is rendered according to \href{https://api.slack.com/reference/surfaces/formatting#basics}{mrkdwn formatting} or not. Defaults to true.}
}
\value{
A message object
}
\description{
Views are app-customized visual areas within modals and Home tabs.
}
\seealso{
\url{https://api.slack.com/reference/messaging/payload}

Other Messaging: 
\code{\link{attachment_field_object}()},
\code{\link{attachment_object}()}
}
\concept{Messaging}
