% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_objects.R
\name{option_group}
\alias{option_group}
\title{Options Group Object}
\usage{
option_group(label, options)
}
\arguments{
\item{label}{A plain_text only \code{\link{text_object}} that defines the label shown above this group of options. Maximum length for the text in this field is 75 characters.}

\item{options}{A list of \code{\link{option_objects}} that belong to this specific group. Maximum of 100 items.}
}
\description{
Provides a way to group options in a select menu or multi-select menu.
}
\seealso{
https://api.slack.com/reference/block-kit/composition-objects#option_group
}
