% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_objects.R
\name{option_object}
\alias{option_object}
\title{Option Object}
\usage{
option_object(text, value, description = NULL, url = NULL)
}
\arguments{
\item{text}{A \code{\link{text_object}} that defines the text shown in the option on the menu. Overflow, select, and multi-select menus can only use plain_text objects, while radio buttons and checkboxes can use mrkdwn text objects. Maximum length for the text in this field is 75 characters.}

\item{value}{The string value that will be passed to your app when this option is chosen. Maximum length for this field is 75 characters.}

\item{description}{A plain_text only text object that defines a line of descriptive text shown below the text field beside the radio button. Maximum length for the text object within this field is 75 characters.}

\item{url}{A URL to load in the user's browser when the option is clicked. The url attribute is only available in overflow menus. Maximum length for this field is 3000 characters. If you're using url, you'll still receive an interaction payload and will need to send an acknowledgement response.}
}
\description{
An object that represents a single selectable item in a select menu, multi-select menu, checkbox group, radio button group, or overflow menu.
}
\seealso{
https://api.slack.com/reference/block-kit/composition-objects#option

Other Composition: 
\code{\link{confirm_object}()},
\code{\link{conversation_list_filter}()},
\code{\link{dispaction_action_configuration}()},
\code{\link{option_group}()},
\code{\link{text_object}()}
}
\concept{Composition}
