% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_objects.R
\name{option_object_list}
\alias{option_object_list}
\title{Create List of Option Objects}
\usage{
option_object_list(df)
}
\arguments{
\item{df}{Table-like object containing columns text and value, and optionally text_type, value, description, and url. See \code{\link{option_object}} documentation for details.}
}
\description{
Helper functions for making lists of options quickly from a dataframe.
}
