% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversations.R
\name{post_message}
\alias{post_message}
\title{Sends a message to a channel.}
\usage{
post_message(
  token = Sys.getenv("SLACK_TOKEN"),
  channel,
  message = NULL,
  as_user = NULL,
  attachments = NULL,
  blocks = NULL,
  icon_emoji = NULL,
  icon_url = NULL,
  link_names = NULL,
  mrkdwn = NULL,
  parse = NULL,
  reply_broadcast = NULL,
  text = NULL,
  thread_ts = NULL,
  unfurl_links = NULL,
  unfurl_media = NULL,
  username = NULL,
  return_response = F
)
}
\arguments{
\item{token}{Authentication token bearing required scopes. Tokens should be passed as an HTTP Authorization header or alternatively, as a POST parameter.}

\item{channel}{Channel, private group, or IM channel to send message to. Can be an encoded ID, or a name. See below for more details.}

\item{message}{A message object.}

\item{as_user}{Pass true to post the message as the authed user, instead of as a bot. Defaults to false. See authorship below.}

\item{attachments}{A JSON-based array of structured attachments, presented as a URL-encoded string.}

\item{blocks}{A JSON-based array of structured blocks, presented as a URL-encoded string.}

\item{icon_emoji}{Emoji to use as the icon for this message. Overrides icon_url. Must be used in conjunction with as_user set to false, otherwise ignored. See authorship below.}

\item{icon_url}{URL to an image to use as the icon for this message. Must be used in conjunction with as_user set to false, otherwise ignored. See authorship below.}

\item{link_names}{Find and link channel names and usernames.}

\item{mrkdwn}{Disable Slack markup parsing by setting to false. Enabled by default. Default: true}

\item{parse}{Change how messages are treated. Defaults to none. See below.}

\item{reply_broadcast}{Used in conjunction with thread_ts and indicates whether reply should be made visible to everyone in the channel or conversation. Defaults to false.}

\item{text}{How this field works and whether it is required depends on other fields you use in your API call. See below for more detail.}

\item{thread_ts}{Provide another message's ts value to make this message a reply. Avoid using a reply's ts value; use its parent instead.}

\item{unfurl_links}{Pass true to enable unfurling of primarily text-based content.}

\item{unfurl_media}{Pass false to disable unfurling of media content.}

\item{username}{Set your bot's user name. Must be used in conjunction with as_user set to false, otherwise ignored. See authorship below.}

\item{return_response}{Whether or not to return the API call response as opposed to the response body. Defaults to FALSE (return response body)}

\item{container_id}{}

\item{file_annotation}{}
}
\description{
Sends a message to a channel.
}
\section{Details}{

This method supports application/json via HTTP POST. Present your token in your request's Authorization header. Learn more.
}

