% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_elements.R
\name{radio_button_group_element}
\alias{radio_button_group_element}
\title{Radio Button Group Element}
\usage{
radio_button_group_element(
  action_id,
  options,
  initial_option = NULL,
  confirm = NULL
)
}
\arguments{
\item{action_id}{An identifier for the action triggered when the checkbox group is changed. You can use this when you receive an interaction payload to \href{https://api.slack.com/interactivity/handling#payloads}{identify the source of the action}. Should be unique among all other action_ids in the containing block. Maximum length for this field is 255 characters.}

\item{options}{A list of \code{\link{option_object}}s. A maximum of 10 options are allowed.}

\item{confirm}{A \code{\link{confirm_object}} that defines an optional confirmation dialog that appears after clicking one of the radio buttons in this element.}

\item{initital_option}{An \code{\link{option_object}} that exactly matches one of the options within options. This option will be selected when the radio button group initially loads.}
}
\description{
A radio button group that allows a user to choose one item from a list of possible options. To use interactive components, you will need to make some changes to prepare your app. Read our \href{https://api.slack.com/interactivity/handling}{guide to enabling interactivity}. Radio buttons are supported in the following app surfaces: Home tabs Modals Messages. Works with block types: Section Actions Input
}
\seealso{
\url{https://api.slack.com/reference/block-kit/block-elements#radio}

Other Elements: 
\code{\link{button_element}()},
\code{\link{channels_multi_select_menu}()},
\code{\link{checkbox_element}()},
\code{\link{conversations_multi_select_menu}()},
\code{\link{datepicker_element}()},
\code{\link{external_multi_select_menu}()},
\code{\link{image_element}()},
\code{\link{overflow_menu_element}()},
\code{\link{plain_text_input_element}()},
\code{\link{static_multi_select_menu}()},
\code{\link{users_multi_select_menu}()}
}
\concept{Elements}
