% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_elements.R
\name{static_select_menu}
\alias{static_select_menu}
\title{Static Options Select Menu Element}
\usage{
static_select_menu(
  placeholder,
  action_id,
  options = NULL,
  option_groups = NULL,
  initial_option = NULL,
  confirm = NULL
)
}
\arguments{
\item{placeholder}{A plain_text only \code{\link{text_object}} that defines the placeholder text shown on the menu. Maximum length for the text in this field is 150 characters.}

\item{action_id}{An identifier for the action triggered when a menu option is selected. You can use this when you receive an interaction payload to \href{https://api.slack.com/messaging/interactivity/enabling#understanding_payloads}{identify the source of the action}. Should be unique among all other action_ids in the containing block. Maximum length for this field is 255 characters.}

\item{options}{A list of \code{\link{option_object}}s. Maximum number of options is 100. If option_groups is specified, this field should not be.}

\item{option_groups}{A list of \code{\link{option_group}} objects. Maximum number of option groups is 100. If options is specified, this field should not be.}

\item{initial_option}{A single \code{link{option_object}}s that exactly match one or more of the options within option_objects or option_groups. These options will be selected when the menu initially loads.}

\item{confirm}{A \code{\link{confirm_object}} that defines an optional confirmation dialog that appears after a menu item is selected.}
}
\description{
This is the simplest form of select menu, with a static list of options passed in when defining the element. A select menu, just as with a standard HTML \verb{<select>} tag, creates a drop down menu with a list of options for a user to choose. The select menu also includes type-ahead functionality, where a user can type a part or all of an option string to filter the list. To use interactive components, you will need to make some changes to prepare your app. Read our \href{https://api.slack.com/interactivity/handling}{guide to enabling interactivity}. Works with block types: Section Actions Input
}
\seealso{
\url{https://api.slack.com/reference/block-kit/block-elements#static_select}

Other Elements: 
\code{\link{button_element}()},
\code{\link{channels_multi_select_menu}()},
\code{\link{channels_select_menu}()},
\code{\link{checkbox_element}()},
\code{\link{conversations_multi_select_menu}()},
\code{\link{conversations_select_menu}()},
\code{\link{datepicker_element}()},
\code{\link{external_multi_select_menu}()},
\code{\link{external_select_menu}()},
\code{\link{image_element}()},
\code{\link{overflow_menu_element}()},
\code{\link{plain_text_input_element}()},
\code{\link{radio_button_group_element}()},
\code{\link{static_multi_select_menu}()},
\code{\link{timepicker_element}()},
\code{\link{users_multi_select_menu}()},
\code{\link{users_select_menu}()}
}
\concept{Elements}
