% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_objects.R
\name{text.object}
\alias{text.object}
\title{Text object}
\usage{
\method{text}{object}(type, text, emoji = NULL, verbatim = NULL)
}
\arguments{
\item{type}{The formatting to use for this text object. Can be one of plain_text or mrkdwn.}

\item{text}{The text for the block. This field accepts any of the standard text formatting markup when type is mrkdwn.}

\item{emoji}{Indicates whether emojis in a text field should be escaped into the colon emoji format. This field is only usable when type is plain_text.}

\item{verbatim}{When set to false (as is default) URLs will be auto-converted into links, conversation names will be link-ified, and certain mentions will be automatically parsed. Using a value of true will skip any preprocessing of this nature, although you can still include manual parsing strings. This field is only usable when type is mrkdwn.}
}
\value{
A Slack Text Object
}
\description{
An object containing some text, formatted either as plain_text or using mrkdwn, our proprietary contribution to the much beloved Markdown standard.
}
\seealso{
\url{https://api.slack.com/reference/block-kit/composition-objects#text}
}
