% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_elements.R
\name{timepicker_element}
\alias{timepicker_element}
\title{Time Picker Element}
\usage{
timepicker_element(
  action_id,
  placeholder = NULL,
  initial_time = NULL,
  confirm = NULL
)
}
\arguments{
\item{action_id}{An identifier for the action triggered when a menu option is selected. You can use this when you receive an interaction payload to \href{https://api.slack.com/messaging/interactivity/enabling#understanding_payloads}{identify the source of the action}. Should be unique among all other action_ids in the containing block. Maximum length for this field is 255 characters.}

\item{placeholder}{A plain_text only \code{\link{text_object}} that defines the placeholder text shown on the timepicker Maximum length for the text in this field is 150 characters.}

\item{initial_time}{The initial time that is selected when the element is loaded. This should be in the format HH:mm, where HH is the 24-hour format of an hour (00 to 23) and mm is minutes with leading zeros (00 to 59), for example 22:25 for 10:25pm.}

\item{confirm}{A \code{\link{confirm_object}} that defines an optional confirmation dialog that appears before the select choices are submitted.}
}
\description{
An element which allows selection of a time of day. On desktop clients, this time picker will take the form of a dropdown list with free-text entry for precise choices. On mobile clients, the time picker will use native time picker UIs. To use interactive components, you will need to make some changes to prepare your app. Read our \href{https://api.slack.com/interactivity/handling}{guide to enabling interactivity}. Works with block types: Section Actions Input
}
\seealso{
\url{https://api.slack.com/reference/block-kit/block-elements#timepicker}

Other Elements: 
\code{\link{button_element}()},
\code{\link{channels_multi_select_menu}()},
\code{\link{channels_select_menu}()},
\code{\link{checkbox_element}()},
\code{\link{conversations_multi_select_menu}()},
\code{\link{conversations_select_menu}()},
\code{\link{datepicker_element}()},
\code{\link{external_multi_select_menu}()},
\code{\link{external_select_menu}()},
\code{\link{image_element}()},
\code{\link{overflow_menu_element}()},
\code{\link{plain_text_input_element}()},
\code{\link{radio_button_group_element}()},
\code{\link{static_multi_select_menu}()},
\code{\link{static_select_menu}()},
\code{\link{users_multi_select_menu}()},
\code{\link{users_select_menu}()}
}
\concept{Elements}
