% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{users_info}
\alias{users_info}
\title{Gets information about a user.}
\usage{
users_info(
  token = Sys.getenv("SLACK_TOKEN"),
  user,
  include_locale = NULL,
  return_response = F
)
}
\arguments{
\item{token}{Authentication token bearing required scopes. Tokens should be passed as an HTTP Authorization header or alternatively, as a POST parameter.}

\item{user}{User to get info on.}

\item{include_locale}{Set this to TRUE to receive the locale for this user. Defaults to FALSE.}

\item{return_response}{Whether or not to return the API call response as opposed to the response body. Defaults to FALSE (return response body)}
}
\value{
A list of channels
}
\description{
Gets information about a user.
}
\seealso{
\url{https://api.slack.com/methods/users.info}

Other users: 
\code{\link{users_list}()}
}
\concept{users}
