% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{verify_request}
\alias{verify_request}
\title{Verify Request}
\usage{
verify_request(
  request_timestamp,
  request_signature,
  signing_secret,
  version = "v0"
)
}
\arguments{
\item{request_timestamp}{The X-Slack-Request-Timestamp header on the HTTP request.}

\item{request_signature}{The X-Slack-Signature header on the request.}

\item{signing_secret}{Your app's \href{https://api.slack.com/authentication/verifying-requests-from-slack#signing_secrets_admin_page}{signing secret} available in the app admin panel under Basic Info.}

\item{version}{The version number right now is always v0.}
}
\value{
TRUE upon successful verification, otherwise an error is thrown.
}
\description{
With the help of signing secrets, your app can more confidently verify whether requests from Slack are authentic.
}
\seealso{
\url{https://api.slack.com/authentication/verifying-requests-from-slack}
}
