% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{view_object}
\alias{view_object}
\title{View Object}
\usage{
view_object(
  type,
  title,
  blocks,
  close = NULL,
  submit = NULL,
  private_metadata = NULL,
  callback_id = NULL,
  clear_on_close = NULL,
  notify_on_close = NULL,
  external_id = NULL
)
}
\arguments{
\item{type}{The type of view. Set to modal for modals and home for Home tabs. Used for Modals and Home Tabs.}

\item{title}{The title that appears in the top-left of the modal. Must be a plain_text text element with a max length of 24 characters. Used for Modals.}

\item{blocks}{A list of blocks that defines the content of the view. Max of 100 blocks. Used for Modals and Home tabs.}

\item{close}{An optional \code{\link{text_object}} that defines the text displayed in the close button at the bottom-right of the view. Max length of 24 characters. Used for Modals.}

\item{submit}{An optional \code{\link{text_object}} that defines the text displayed in the submit button at the bottom-right of the view. submit is required when an input block is within the blocks array. Max length of 24 characters. Used for Modals.}

\item{private_metadata}{An optional string that will be sent to your app in view_submission and block_actions events. Max length of 3000 characters. Used for Modals Home tabs.}

\item{callback_id}{An identifier to recognize interactions and submissions of this particular view. Don't use this to store sensitive information (use private_metadata instead). Max length of 255 characters. Used for Modals Home tabs.}

\item{clear_on_close}{When set to true, clicking on the close button will clear all views in a modal and close it. Defaults to false. Used for Modals.}

\item{notify_on_close}{Indicates whether Slack will send your request URL a view_closed event when a user clicks the close button. Defaults to false. Used for Modals.}

\item{external_id}{A custom identifier that must be unique for all views on a per-team basis. Used for Modals Home tabs.}
}
\value{
A View Object
}
\description{
Views are app-customized visual areas within modals and Home tabs.
}
\seealso{
\url{https://api.slack.com/reference/surfaces/views}

Other Views: 
\code{\link{views_open}()},
\code{\link{views_push}()},
\code{\link{views_update}()}
}
\concept{Views}
