% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{views.open}
\alias{views.open}
\title{Open a View}
\usage{
views.open(token, trigger_id, view)
}
\arguments{
\item{token}{Authentication token bearing required scopes. Tokens should be passed as an HTTP Authorization header or alternatively, as a POST parameter.}

\item{trigger_id}{Exchange a trigger to post to the user.}

\item{view}{A \code{\link{view.object}}. See \url{https://api.slack.com/reference/surfaces/views} for more details.}
}
\value{
A \code{\link{view.object}} with status or an error message.
}
\description{
Open a view for a user.
}
\seealso{
\url{https://api.slack.com/methods/views.open}
}
