% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{views.push}
\alias{views.push}
\title{Push a View}
\usage{
views.push(token, trigger_id, view)
}
\arguments{
\item{token}{Authentication token bearing required scopes. Tokens should be passed as an HTTP Authorization header or alternatively, as a POST parameter.}

\item{trigger_id}{Exchange a trigger to post to the user.}

\item{view}{A \code{\link{view_object}}.}
}
\value{
If you pass a valid view object along with a valid trigger_id, you'll receive a success response with the view object that was pushed to the stack.
}
\description{
Push a new view onto the existing view stack by passing a view object and a valid trigger_id generated from an interaction within the existing modal. The pushed view is added to the top of the stack, so the user will go back to the previous view after they complete or cancel the pushed view. After a modal is opened, the app is limited to pushing 2 additional views. Read the \href{https://api.slack.com/surfaces/modals}{modals documentation} to learn more about the lifecycle and intricacies of views.
}
\seealso{
\url{https://api.slack.com/methods/views.push}
}
