% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{views_open}
\alias{views_open}
\title{Open a View for a User}
\usage{
views_open(token, trigger_id, view, return_response = F)
}
\arguments{
\item{token}{Authentication token bearing required scopes. Tokens should be passed as an HTTP Authorization header or alternatively, as a POST parameter.}

\item{trigger_id}{Exchange a trigger to post to the user.}

\item{view}{A \code{\link{view_object}}.}
}
\value{
A \code{\link{view_object}} with status or an error message.
}
\description{
Open a modal with a user by exchanging a trigger_id received from another interaction. See the \href{https://api.slack.com/surfaces/modals}{modals documentation} to learn how to obtain triggers from interactive components.
}
\seealso{
\url{https://api.slack.com/methods/views.open}

Other Views: 
\code{\link{view_object}()},
\code{\link{views_push}()},
\code{\link{views_update}()}
}
\concept{Views}
