% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{views_publish}
\alias{views_publish}
\title{Publish a View}
\usage{
views_publish(token, user_id, view, hash = NULL, return_response = F)
}
\arguments{
\item{token}{Authentication token bearing required scopes. Tokens should be passed as an HTTP Authorization header or alternatively, as a POST parameter.}

\item{user_id}{id of the user you want publish a view to.}

\item{view}{A \code{\link{view_object}}.}

\item{hash}{A string that represents view state to protect against possible race conditions.}

\item{return_response}{Whether or not to return the API call response as opposed to the response body. Defaults to FALSE (return response body)}
}
\value{
Assuming your view object was properly formatted, valid, and the user_id was viable, you will receive a success response.
}
\description{
Create or update the view that comprises an app's Home tab for a specific user.
}
\seealso{
\url{https://api.slack.com/methods/views.publish}

Other Views: 
\code{\link{view_object}()},
\code{\link{views_open}()},
\code{\link{views_push}()},
\code{\link{views_update}()}
}
\concept{Views}
