% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{views_push}
\alias{views_push}
\title{Push a View}
\usage{
views_push(
  token = Sys.getenv("SLACK_TOKEN"),
  trigger_id,
  view,
  return_response = F
)
}
\arguments{
\item{token}{Authentication token bearing required scopes. Tokens should be passed as an HTTP Authorization header or alternatively, as a POST parameter.}

\item{trigger_id}{Exchange a trigger to post to the user.}

\item{view}{A \code{\link{view_object}}.}

\item{return_response}{Whether or not to return the API call response as opposed to the response body. Defaults to FALSE (return response body)}
}
\value{
If you pass a valid view object along with a valid trigger_id, you'll receive a success response with the view object that was pushed to the stack.
}
\description{
Push a new view onto the existing view stack by passing a view object and a valid trigger_id generated from an interaction within the existing modal. The pushed view is added to the top of the stack, so the user will go back to the previous view after they complete or cancel the pushed view. After a modal is opened, the app is limited to pushing 2 additional views. Read the \href{https://api.slack.com/surfaces/modals}{modals documentation} to learn more about the lifecycle and intricacies of views.
}
\seealso{
\url{https://api.slack.com/methods/views.push}

Other Views: 
\code{\link{view_object}()},
\code{\link{views_open}()},
\code{\link{views_publish}()},
\code{\link{views_update}()}
}
\concept{Views}
