% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_elements.R
\name{image_element}
\alias{image_element}
\title{Image Element}
\usage{
image_element(image_url, alt_text)
}
\arguments{
\item{image_url}{The URL of the image to be displayed.}

\item{alt_text}{A plain-text summary of the image. This should not contain any markup.}
}
\description{
An element to insert an image as part of a larger block of content. If you want a block with only an image in it, you're looking for the image block. Works with block types: Section Context
}
\seealso{
\url{https://api.slack.com/reference/block-kit/block-elements#image}

Other Elements: 
\code{\link{button_element}()},
\code{\link{channels_multi_select_menu}()},
\code{\link{channels_select_menu}()},
\code{\link{checkbox_element}()},
\code{\link{conversations_multi_select_menu}()},
\code{\link{conversations_select_menu}()},
\code{\link{datepicker_element}()},
\code{\link{external_multi_select_menu}()},
\code{\link{external_select_menu}()},
\code{\link{overflow_menu_element}()},
\code{\link{plain_text_input_element}()},
\code{\link{radio_button_group_element}()},
\code{\link{static_multi_select_menu}()},
\code{\link{static_select_menu}()},
\code{\link{timepicker_element}()},
\code{\link{users_multi_select_menu}()},
\code{\link{users_select_menu}()}
}
\concept{Elements}
